FootnoteMagic: refactor reflowing.
Footnote bubbles are now reflowed just before they are shown, if
necessary. They also reflow if they are visible when the window size
changes, otherwise their reflow is deferred to when they are being
shown. In theory, this results in the minimal amount of computation,
especially in the case that no footnotes bubbles are shown.